requirements.txt
fastapi==0.111.1
pydantic==2.8.2
PyJWT==2.9.0
starlette==0.38.2
uvicorn==0.30.6
PyMySQL~=1.1.1
PyYAML==6.0.2
setuptools==72.2.0
loguru==0.7.2
schedule==1.2.2
websockets==12.0